+++ src/Credfeto.DotNet.Repo.Tools.TemplateUpdate/Services/BulkTemplateUpdater.cs
@@ -14,6 +14,7 @@ using Credfeto.DotNet.Repo.Tools.Build.Interfaces;
 using Credfeto.DotNet.Repo.Tools.Build.Interfaces.Exceptions;
 using Credfeto.DotNet.Repo.Tools.DotNet.Interfaces;
 using Credfeto.DotNet.Repo.Tools.Git.Interfaces;
+using Credfeto.DotNet.Repo.Tools.Git.Interfaces.Exceptions;
 using Credfeto.DotNet.Repo.Tools.Models;
 using Credfeto.DotNet.Repo.Tools.Models.Packages;
 using Credfeto.DotNet.Repo.Tools.Packages.Interfaces;
@@ -140,6 +141,11 @@ public sealed class BulkTemplateUpdater : IBulkTemplateUpdater
                 {
                     this._logger.LogBuildFailedOnCreateRelease(message: exception.Message, exception: exception);
                 }
+                catch (GitRepositoryLockedException exception)
+                {
+                    this._logger.LogRepoLocked(repo, exception.Message,exception: exception);
+                }
+
                 finally
                 {
                     if (!string.IsNullOrWhiteSpace(updateContext.TrackingFileName))